// frontend/src/api/generated/models/ProblemType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ProblemType {
    POTHOLE = 'pothole',
    GARBAGE = 'garbage',
    ROAD_WORK = 'road_work',
    POLLUTION = 'pollution',
    TRAFFIC_LIGHT = 'traffic_light',
    FLOODING = 'flooding',
    LIGHTING = 'lighting',
    CONSTRUCTION = 'construction',
    ROADS = 'roads',
    INFRASTRUCTURE = 'infrastructure',
    OTHER = 'other',
}